+++ EmpiriaBMS.Core/Dtos/SupportiveWorkDto.cs
@@ -5,13 +5,13 @@ namespace EmpiriaBMS.Core.Dtos;
 
 public class SupportiveWorkDto : EntityDto
 {
-    public int TypeId { get; set; }
+    public long TypeId { get; set; }
     public SupportiveWorkType Type { get; set; }
 
     public long MenHours { get; set; }
 
     public float CompletionEstimation { get; set; }
 
-    public int DisciplineId { get; set; }
+    public long DisciplineId { get; set; }
     public Discipline Discipline { get; set; }
 }
