+++ EmpiriaBMS.Models/Models/Email.cs
@@ -8,12 +8,12 @@ public class Email : Entity
     [Required]
     public string? Address { get; set; }
 
-    public int? UserId { get; set; }
+    public long? UserId { get; set; }
     public User? User { get; set; }
 
-    public int? ClientId { get; set; }
+    public long? ClientId { get; set; }
     public Client? Client { get; set; }
 
-    public int? SubConstructorId { get; set; }
+    public long? SubConstructorId { get; set; }
     public SubConstructor? SubConstructor { get; set; }
 }
